media: msm: camera_v2: jpeg: Fix jpeg max event queue timeout
At some point the max event queue timeout was set to UINT_MAX,
which for us is 0xFFFFFFFF, but this is absolutely wrong,
as the Linux scheduler's maximum scheduling timeout is
LONG_MAX, which is 0x7FFFFFFF.
Waiting for more time than the maximum scheduling timeout will
lead to major unstabilities, such as sched crashes, watchdog
firing and who knows what else!

Set it back to the proper value of LONG_MAX to solve these
possible issues.